Whoopsies

+++ src/Utils/Utils.cpp
@@ -152,7 +152,7 @@ namespace Utils
 
 	float Vec3SqrDistance(float(*v1)[3], float(*v2)[3])
 	{
-		return (std::pow((*v2)[0] - (*v1)[0], 2) + std::pow((*v2)[1] - (*v1)[1], 2) + std::pow((*v2)[2] - (*v1)[2], 2));
+		return static_cast<float>(std::pow((*v2)[0] - (*v1)[0], 2) + std::pow((*v2)[1] - (*v1)[1], 2) + std::pow((*v2)[2] - (*v1)[2], 2));
 	}
